Actually Fix rspec-vim
Terms out the term command will run in place and once finished will
remove itself. This causes the file you are running to disappear if you
are using a split.

Forcing the terminal to split before running the term command opens a
new terminal where all the action can happen.