Remove termencoding option from vim config
Removse the termencoding configuration option from the neovim
configuration file in lazy.vim due to this option breaking the editor as
the property does not exist.

I could not find much in the changelog for Neovim with this option and
the LSP is still picking up the option so unsure if this is a bug in the
release of Neovim or not.